Add pandasdmx.exceptions to unify exception classes
